requirements.txt
pymeshlab
bs4
twitchio
obs-websocket-py
numpy
retry
python-thingiverse
